requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.1
fastapi==0.110.1
fastjsonschema==2.20.0
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_metadata==8.0.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
Mako==1.3.2
MarkupSafe==2.1.5
more-itertools==10.3.0
msgpack==1.0.8
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32-ctypes==0.2.2
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tomlkit==0.12.5
trove-classifiers==2024.5.22
typing_extensions==4.11.0
urllib3==2.2.2
uvicorn==0.29.0
virtualenv==20.26.3
zipp==3.19.2